#9 Token not returning
This issue is caused by `xerotoken.json` not existing by default and `StoreToken` only persisting the returned token if it does. Since `File.WriteAllText` creates the file if it doesn't already exist or overwrites it if it does, removing the `File.Exists` check means that it will create/update the file as expected and avoid the issue of a missing `refresh_token`.